+++ src/main/java/com/en_circle/slt/plugin/lisp/psi/impl/LispPsiImplUtil.java
@@ -117,17 +117,17 @@ public class LispPsiImplUtil {
         return element;
     }
 
-    private static LispSymbol createSymbol(Project project, String name) {
+    public static LispSymbol createSymbol(Project project, String name) {
         LispFile file = createFile(project, name);
         return PsiTreeUtil.findChildOfType(file, LispSymbol.class);
     }
 
-    private static LispComment createComment(Project project, String name) {
+    public static LispComment createComment(Project project, String name) {
         LispFile file = createFile(project, name);
         return PsiTreeUtil.findChildOfType(file, LispComment.class);
     }
 
-    private static LispString createString(Project project, String name) {
+    public static LispString createString(Project project, String name) {
         LispFile file = createFile(project, name);
         return PsiTreeUtil.findChildOfType(file, LispString.class);
     }